+++ datumaro/cli/contexts/model.py
@@ -1,5 +1,4 @@
-
-# Copyright (C) 2019-2020 Intel Corporation
+# Copyright (C) 2019-2021 Intel Corporation
 #
 # SPDX-License-Identifier: MIT
 
@@ -7,13 +6,11 @@ import argparse
 import logging as log
 import os
 import os.path as osp
-import re
 
-from datumaro.components.dataset import DEFAULT_FORMAT
 from datumaro.components.project import Environment
 
-from ...util import CliException, MultilineFormatter, add_subparser
-from ...util.project import load_project, \
+from ..util import CliException, MultilineFormatter, add_subparser
+from ..util.project import load_project, \
     generate_next_name, generate_next_file_name
